+++ src/kiara/operations/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import typing
 
-from kiara.module_config import OperationConfig
+from kiara.module_config import ModuleInstanceConfig
 
 if typing.TYPE_CHECKING:
     from kiara import Kiara, KiaraModule
@@ -16,7 +16,7 @@ class ClassAttributes(object):
 class Operation(object):
     def create(
         self,
-        config: typing.Union[OperationConfig, typing.Mapping, str],
+        config: typing.Union[ModuleInstanceConfig, typing.Mapping, str],
         module_config: typing.Union[
             None, typing.Mapping[str, typing.Any], "PipelineModuleConfig"
         ] = None,
@@ -28,13 +28,15 @@ class Operation(object):
 
             kiara = Kiara.instance()
 
-        operation_config = OperationConfig.create(
+        operation_config = ModuleInstanceConfig.create(
             config=config, module_config=module_config, kiara=kiara
         )
 
         return Operation(config=operation_config, kiara=kiara)
 
-    def __init__(self, config: OperationConfig, kiara: typing.Optional["Kiara"] = None):
+    def __init__(
+        self, config: ModuleInstanceConfig, kiara: typing.Optional["Kiara"] = None
+    ):
 
         if kiara is None:
             from kiara import Kiara
@@ -42,7 +44,7 @@ class Operation(object):
             kiara = Kiara.instance()
 
         self._kiara: Kiara = kiara
-        self._config: OperationConfig = config
+        self._config: ModuleInstanceConfig = config
         self._module: typing.Optional["KiaraModule"] = None
 
     def module(self) -> "KiaraModule":